fix(wayland): app_pids() returns the sway process subtree for a11y (M11b)
Twin of M11: glass-wayland had no app_pid/app_pids override, so ctx.pids was
always empty and the AT-SPI reader fell back to 'accept the first app on the
bus' — fine with one app, mis-targets with several. The app is a sway 'exec'
descendant (shell, and bwrap when sandboxed), so app_pids() now returns
proc_tree_pids(sway_child.id()).

Hoist proc_tree_pids + collect_descendants out of glass-x11 into a new
focused glass-proc-linux crate (generic /proc tree walk — not sandbox logic,
and OS-specific so not glass-core) shared by both Linux backends; x11 now
imports it (4 tests moved with the code). The Windows peer descendant_pids
stays in glass-windows (different OS API).
