+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["crates/glass-core", "crates/glass-x11", "crates/glass-testapp", "crates/glass-mcp", "crates/glass-wayland", "crates/glass-a11y-linux", "crates/glass-windows", "crates/glass-a11y-windows", "crates/glass-sandbox-linux", "crates/glass-clip-hook"]
+members = ["crates/glass-core", "crates/glass-x11", "crates/glass-testapp", "crates/glass-mcp", "crates/glass-wayland", "crates/glass-a11y-linux", "crates/glass-windows", "crates/glass-a11y-windows", "crates/glass-sandbox-linux", "crates/glass-proc-linux", "crates/glass-clip-hook"]
 
 [workspace.package]
 edition = "2021"
@@ -19,6 +19,7 @@ tempfile = "3"
 x11rb = { version = "0.13", features = ["xtest"] }
 glass-core = { path = "crates/glass-core" }
 glass-sandbox-linux = { path = "crates/glass-sandbox-linux" }
+glass-proc-linux = { path = "crates/glass-proc-linux" }
 glass-wayland = { path = "crates/glass-wayland" }
 glass-a11y-linux = { path = "crates/glass-a11y-linux" }
 atspi = { version = "0.30", features = ["tokio"] }
