+++ crates/glass-x11/Cargo.toml
@@ -14,6 +14,7 @@ bench = false
 [dependencies]
 glass-core.workspace = true
 glass-sandbox-linux.workspace = true
+glass-proc-linux.workspace = true
 x11rb.workspace = true
 
 [dev-dependencies]
